Overhaul routes and menus
