--- src/pages/Home/components/Header.test.tsx
@@ -1,7 +1,0 @@
-import { screen } from '@testing-library/react'
-
-describe('HomeHeader', () => {
-  it('renders correctly', async () => {
-    expect(screen.getByTestId('homeHeader')).toBeInTheDocument()
-  })
-})
