+++ src/pages/LoadUnload/LoadUnloadPage.tsx
@@ -12,14 +12,12 @@ import styled from 'styled-components'
 import 'leaflet/dist/leaflet.css'
 
 import { HEADER_HEIGHT, Z_INDEX_MODAL } from '../../shared/constants'
-import FeedbackModal from '../../shared/components/FeedbackModal'
 import { MainContent, PageWrapper } from '../../shared/components/FullPageSize'
 import { MapStyle } from '../../shared/map/mapStyle'
 import { defaultMapOptions, setMapDefaults } from '../../shared/map/mapDefaults'
 import { useDocumentTitle } from '../../shared/hooks/useDocumentTitle'
 
 import { LoadUnloadAddressForm } from './components/AddressForm'
-import { LoadUnloadHeader } from './components/Header'
 import { LoadUnloadDetailFeature } from './components/DetailFeature'
 import { LoadUnloadMapLayers } from './components/MapLayers'
 import { ModalDateTime } from './components/ModalDateTime'
@@ -27,6 +25,7 @@ import { LoadUnloadViewerContainer } from './components/ViewerContainer'
 import { LoadUnloadMapSettingsDisplay } from './components/MapSettingsDisplay'
 import { LoadUnloadMapProvider } from './contexts/MapProvider'
 import { LoadUnloadPageProvider } from './contexts/PageProvider'
+import { Header } from '../../shared/components/Header'
 
 const { SnapPoint } = mapPanelConstants
 
@@ -54,16 +53,12 @@ const LoadUnloadPage = () => {
 
   const [showAddressForm, setShowAddressForm] = useState(false)
   const [showDateTimeModal, setShowDateTimeModal] = useState(false)
-  const [openFeedbackModal, setOpenFeedbackModal] = useState(false)
   const Element = showDesktopVariant ? MapPanel : StyledMapPanelDrawer
 
   return (
     <LoadUnloadPageProvider>
       <PageWrapper>
-        <LoadUnloadHeader
-          setOpenFeedbackModal={setOpenFeedbackModal}
-          title="Laden en lossen"
-        />
+        <Header title="Laden en lossen" />
 
         <MainContent data-testid="load-unload-page">
           <MapStyle />
@@ -108,8 +103,6 @@ const LoadUnloadPage = () => {
         showModal={showDateTimeModal}
         setShowModal={setShowDateTimeModal}
       />
-
-      <FeedbackModal setOpen={setOpenFeedbackModal} open={openFeedbackModal} />
     </LoadUnloadPageProvider>
   )
 }
