+++ src/pages/ProhibitorySigns/ProhibitorySignsPage.tsx
@@ -12,12 +12,10 @@ import styled from 'styled-components'
 import 'leaflet/dist/leaflet.css'
 
 import { HEADER_HEIGHT } from '../../shared/constants'
-import FeedbackModal from '../../shared/components/FeedbackModal'
 import { MainContent, PageWrapper } from '../../shared/components/FullPageSize'
 import { MapStyle } from '../../shared/map/mapStyle'
 import { defaultMapOptions, setMapDefaults } from '../../shared/map/mapDefaults'
 
-import ProhibitorySignsHeader from './components/Header'
 import ProhibitorySignsDetailFeature from './components/DetailFeature'
 import ProhibitorySignsMapLayers from './components/MapLayers'
 import ProhibitorySignsScenarioWizard from './components/ScenarioWizard'
@@ -25,6 +23,7 @@ import ProhibitorySignsViewerContainer from './components/ViewerContainer'
 import ProhibitorySignsPageProvider from './contexts/PageProvider'
 import ScenarioDisplay from './components/ScenarioDisplay'
 import ProhibitorySignsMapProvider from './contexts/MapProvider'
+import { Header } from '../../shared/components/Header'
 
 const { SnapPoint } = mapPanelConstants
 
@@ -39,9 +38,6 @@ const StyledMapPanelDrawer = styled(MapPanelDrawer)`
 `
 
 const ProhibitorySignsPage = () => {
-  const [openFeedbackModal, setOpenFeedbackModal] = useState(false)
-
-  // variables concerning the map
   const [mapInstance, setMapInstance] = useState<L.Map | null>(null)
   const [showDesktopVariant] = useMatchMedia({ minBreakpoint: 'tabletM' })
 
@@ -56,10 +52,7 @@ const ProhibitorySignsPage = () => {
   return (
     <PageWrapper>
       <ProhibitorySignsPageProvider>
-        <ProhibitorySignsHeader
-          setOpenFeedbackModal={setOpenFeedbackModal}
-          title="Bereikbaarheid op kenteken"
-        />
+        <Header title="Bereikbaarheid op kenteken" />
 
         <MainContent data-testid="prohibitory-signs-page">
           <MapStyle />
@@ -84,12 +77,8 @@ const ProhibitorySignsPage = () => {
           </StyledMap>
         </MainContent>
 
-        <ProhibitorySignsScenarioWizard
-          setShowFeedbackModal={setOpenFeedbackModal}
-        />
+        <ProhibitorySignsScenarioWizard />
       </ProhibitorySignsPageProvider>
-
-      <FeedbackModal setOpen={setOpenFeedbackModal} open={openFeedbackModal} />
     </PageWrapper>
   )
 }
