+++ src/shared/components/Header/HeaderMenuItems.tsx
@@ -7,10 +7,17 @@
 // however TS does not offer a block-scoped ignore at the time of writing
 // see https://github.com/Microsoft/TypeScript/issues/19573
 
-import { Icon, MenuButton, MenuItem } from '@amsterdam/asc-ui'
+import { Icon, MenuButton, MenuItem, themeColor } from '@amsterdam/asc-ui'
 import { generatePath, Link, matchPath, useLocation } from 'react-router-dom'
 import { RouteIds, getPathTo } from '../../../routes'
 import { ExternalLink } from '@amsterdam/asc-assets'
+import styled from 'styled-components'
+
+const MenuDivider = styled.hr`
+  height: 2rem;
+  margin-inline: 1em;
+  border: 1px solid ${themeColor('tint', 'level3')};
+`
 
 export const HeaderMenuItems = () => {
   const location = useLocation()
@@ -19,7 +26,7 @@ export const HeaderMenuItems = () => {
     <>
       <MenuItem>
         <MenuButton
-          as={Link}
+          forwardedAs={Link}
           to={generatePath(getPathTo(RouteIds.LICENCE_PLATE_PAGE))}
           active={
             matchPath(
@@ -36,7 +43,7 @@ export const HeaderMenuItems = () => {
 
       <MenuItem>
         <MenuButton
-          as={Link}
+          forwardedAs={Link}
           to={generatePath(getPathTo(RouteIds.ROAD_OBSTRUCTIONS_PAGE))}
           active={
             matchPath(
@@ -53,7 +60,7 @@ export const HeaderMenuItems = () => {
 
       <MenuItem>
         <MenuButton
-          as={Link}
+          forwardedAs={Link}
           to={generatePath(getPathTo(RouteIds.LOAD_UNLOAD_PAGE))}
           active={
             matchPath(
@@ -70,16 +77,51 @@ export const HeaderMenuItems = () => {
 
       <MenuItem>
         <MenuButton
-          as={Link}
+          forwardedAs={Link}
           to="https://tourbuzz.amsterdam.nl/"
           target="_blank"
           iconRight={
-            <Icon size={16}>
+            <Icon size={13} style={{ marginLeft: -2 }}>
               <ExternalLink />
             </Icon>
           }
         >
-          Tour Buzz
+          Touringcars
+        </MenuButton>
+      </MenuItem>
+
+      <MenuDivider />
+
+      <MenuItem>
+        <MenuButton
+          forwardedAs={Link}
+          to={generatePath(getPathTo(RouteIds.DATA))}
+          active={
+            matchPath(location.pathname, generatePath(getPathTo(RouteIds.DATA)))
+              ? 'true'
+              : undefined
+          }
+          style={{ fontWeight: 'normal' }}
+        >
+          Databronnen
+        </MenuButton>
+      </MenuItem>
+
+      <MenuItem>
+        <MenuButton
+          forwardedAs={Link}
+          to={generatePath(getPathTo(RouteIds.CONTACT))}
+          active={
+            matchPath(
+              location.pathname,
+              generatePath(getPathTo(RouteIds.CONTACT))
+            )
+              ? 'true'
+              : undefined
+          }
+          style={{ fontWeight: 'normal' }}
+        >
+          Contact en hulp
         </MenuButton>
       </MenuItem>
     </>
